+++ client/src/hooks/useMeetVideoState.ts
@@ -1,0 +1,1 @@
+export const useMeetVideoState = () => {};
